# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Tripline API
  version: 0.1.0
  description: "Self-hosted dependency drift monitor. Authenticate with the session cookie (browser; state-changing requests need header `X-Tripline-CSRF: 1`) or a read-only bearer API token."
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: tripline_session
    token:
      type: http
      scheme: bearer
      description: Read-only API token (tl_…)
  schemas:
    def-0:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details: {}
      title: ErrorResponse
paths:
...
  /api/v1/summary:
    get:
      summary: Dashboard counters
      tags:
        - system
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - monitors
                  - openIncidents
                  - openDrift
                  - checksLast24h
                properties:
                  monitors:
                    type: object
                    required:
                      - total
                      - up
                      - down
                      - unknown
                      - paused
                    properties:
                      total:
                        type: integer
                      up:
                        type: integer
                      down:
                        type: integer
                      unknown:
                        type: integer
                      paused:
                        type: integer
                  openIncidents:
                    type: integer
                  openDrift:
                    type: object
                    required:
                      - breaking
                      - warning
                      - info
                    properties:
                      breaking:
                        type: integer
                      warning:
                        type: integer
                      info:
                        type: integer
                  checksLast24h:
                    type: object
                    required:
                      - total
                      - failed
                    properties:
                      total:
                        type: integer
                      failed:
                        type: integer
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/gate:
    get:
      summary: CI deploy gate. Always HTTP 200; check `pass`. Use an API token.
      tags:
        - system
      parameters:
        - schema:
            type: string
            maxLength: 500
          in: query
          name: tags
          required: false
          description: Comma-separated tags; monitors with any of them are evaluated. Empty = all.
        - schema:
            type: string
            maxLength: 100
            default: down,breaking
          in: query
          name: failOn
          required: false
          description: Comma-separated subset of down, unknown, breaking, warning.
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - pass
                  - evaluated
                  - failOn
                  - tags
                  - failures
                  - evaluatedAt
                properties:
                  pass:
                    type: boolean
                  evaluated:
                    type: integer
                  failOn:
                    type: array
                    items:
                      type: string
                  tags:
                    type: array
                    items:
                      type: string
                  failures:
                    type: array
                    items:
                      type: object
                      required:
                        - monitorId
                        - name
                        - reason
                      properties:
                        monitorId:
                          type: string
                          format: uuid
                        name:
                          type: string
                        reason:
                          type: string
                  evaluatedAt:
                    type: string
                    format: date-time
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/audit:
    get:
      summary: Audit log, newest first (cursor pagination by id)
      tags:
        - system
      parameters:
        - schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
          in: query
          name: limit
          required: false
        - schema:
            type: integer
            minimum: 1
          in: query
          name: before
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - entries
                properties:
                  entries:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - action
                        - targetType
                        - targetId
                        - details
                        - ip
                        - createdAt
                        - userId
                        - userEmail
                      properties:
                        id:
                          type: integer
                        action:
                          type: string
                        targetType:
                          anyOf:
                            - type: string
                            - type: "null"
                        targetId:
                          anyOf:
                            - type: string
                            - type: "null"
                        details:
                          type: object
                          additionalProperties: {}
                        ip:
                          anyOf:
                            - type: string
                            - type: "null"
                        createdAt:
                          type: string
                          format: date-time
                        userId:
                          anyOf:
                            - type: string
                              format: uuid
                            - type: "null"
                        userEmail:
                          anyOf:
                            - type: string
                            - type: "null"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
servers:
  - url: /
security:
  - session: []
  - token: []